Update on Swiss corrections
After looking in details to what you have, I think all the final counts are correct. This is due to the fortunate event that 1+2+3+4+5 = 15, and the server was set to retain those keys for 15 days. So the final version of the file was scraped right before it disappeared (gradually, actually!).